+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.4.0] - 2024-x-x
 ### Added
-* Support for v2 of `crwlr/crawler`.
+* Allow usage of `crwlr/crawler` v2.0.
+* Method `StepBuilder::outputType()`, returning a `Crwlr\Crawler\Steps\StepOutputType` enum instance, informing about the possible output type of the underlying step. Currently, there is a default implementation in the abstract `StepBuilder` class, returning `StepOutputType::Mixed`. But this implementation will be removed in v3.0, so child classes should always explicitly define the possible output type. More info in the readme file. __Attention__: As `Crwlr\Crawler\Steps\StepOutputType` was introduced in `crwlr/crawler` v1.8.0, this is now the minimum required version of the crawler package.
 
 ## [2.3.1] - 2024-06-18
 ### Fixed